# cakeshop-api/src/main/webapp/api/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "Cakeshop REST API",
		"version": "1.0.0",
		"contact": {}
	},
	"host": "localhost:8080",
	"schemes": [
		"http"
	],
	"basePath": "/cakeshop/api",
...
	"paths": {
...
		"/contract/compile": {
			"post": {
				"tags": [
					"contract"
				],
				"description": "Compile the given source code into EVM binary",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "code",
						"type": "string",
						"description": "Contract source code",
						"required": true,
						"in": "body"
					},
					{
						"name": "code_type",
						"type": "string",
						"in": "body",
						"description": "Type of code being submitted",
						"required": false,
						"default": "solidity",
						"enum": [
							"solidity"
						]
					},
					{
						"name": "optimize",
						"type": "boolean",
						"required": false,
						"in": "body",
						"description": "Generate optimized code",
						"default": false
					}
				],
				"examples": [
					{
						"name": "Compile SimpleStorage contract",
						"parameters": {
							"body": "{\n\n  \"code\": \"contract SimpleStorage { uint storedData; function\nset(uint x) { storedData = x; } function get() constant returns (uint retVal) { return storedData; } }\",\n\n  \"code_type\": \"solidity\"\n\n}\n"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "n/a for compile"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'contract'"
										},
										"attributes": {
											"$ref": "#/definitions/Contract"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/contract/create": {
			"post": {
				"tags": [
					"contract"
				],
				"description": "Deploy the given contract onto the chain and add it to the Contract Registry",
				"operationId": "postContractCreate",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "code",
						"type": "string",
						"description": "Contract source code",
						"in": "body"
					},
					{
						"name": "code_type",
						"type": "string",
						"description": "Type of code being submitted",
						"required": false,
						"default": "solidity",
						"in": "body",
						"enum": [
							"solidity"
						]
					},
					{
						"name": "args",
						"type": "array",
						"required": false,
						"in": "body",
						"description": "Constructor arguments (if needed)",
						"items": {
							"type": "object"
						}
					},
					{
						"name": "binary",
						"type": "string",
						"required": false,
						"in": "body",
						"description": "If multiple contracts are specified in the source code, a specific binary to deploy can be passed here"
					},
					{
						"name": "optimize",
						"type": "boolean",
						"required": false,
						"default": false,
						"in": "body",
						"description": "Generate optimized code"
					}
				],
				"examples": [
					{
						"name": "Create SimpleStorage contract",
						"parameters": {
							"body": "{\n\n  \"code\": \"contract SimpleStorage { uint storedData; function\nset(uint x) { storedData = x; } function get() constant returns (uint retVal) { return storedData; } }\",\n\n  \"code_type\": \"solidity\"\n\n}\n"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "Transaction ID. Use /transaction/get to get the resulting Contract ID"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'transaction_result'"
										},
										"attributes": {
											"$ref": "#/definitions/TransactionResult"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/contract/get": {
			"post": {
				"tags": [
					"contract"
				],
				"operationId": "postContractGet",
				"description": "Get information about the contract at the given address",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"type": "string",
						"description": "ID (address) of Contract to retrieve",
						"example": "0x01f999deef030a0f3c55cec46e6693c06c6b01f8",
						"in": "body"
					}
				],
				"examples": [
					{
						"name": "Get info about SimpleStorage contract",
						"parameters": {
							"body": "{\n  \"address\": \"<ENTER ADDRESS HERE>\"\n}\n"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "Contract ID"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'contract'"
										},
										"attributes": {
											"$ref": "#/definitions/Contract"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/contract/list": {
			"post": {
				"tags": [
					"contract"
				],
				"description": "List all deployed contracts on the chain",
				"operationId": "postContractList",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "array",
									"description": "response data",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "string",
												"description": "Contract ID"
											},
											"type": {
												"type": "string",
												"description": "resource type identifier, 'contract'"
											},
											"attributes": {
												"$ref": "#/definitions/Contract"
											}
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/contract/read": {
			"post": {
				"tags": [
					"contract"
				],
				"description": "Read contract data using the given method and arguments (does not create a transaction). Optionally read historical state data by passing a block number in the past.",
				"operationId": "postContractRead",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"type": "string",
						"required": true,
						"description": "Contract ID (address hash) to read from",
						"example": "",
						"in": "body"
					},
					{
						"name": "method",
						"type": "string",
						"required": true,
						"description": "Method name to call",
						"in": "body"
					},
					{
						"name": "args",
						"type": "array",
						"required": false,
						"description": "Method arguments (if needed)",
						"in": "body",
						"items": {
							"type": "object"
						}
					},
					{
						"name": "blockNumber",
						"type": "integer",
						"required": false,
						"description": "Block from which to read the state from",
						"in": "body"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"attributes": {
											"properties": {
												"result": {
													"type": "object",
													"description": "return data from the read call"
												}
											}
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
		"/contract/transact": {
			"post": {
				"tags": [
					"contract"
				],
				"description": "Call a transactional method on the given contract. No state will change until the transaction is committed into a Block.",
				"operationId": "postContractTransact",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"type": "string",
						"required": true,
						"description": "Contract ID (address hash) to transact with",
						"example": "",
						"in": "body"
					},
					{
						"name": "method",
						"type": "string",
						"required": true,
						"in": "body",
						"description": "Method name to call"
					},
					{
						"name": "args",
						"type": "array",
						"required": false,
						"description": "Method arguments (if needed)",
						"in": "body",
						"items": {
							"type": "object"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"data": {
									"type": "object",
									"description": "response data",
									"properties": {
										"id": {
											"type": "string",
											"description": "Transaction ID for the operation"
										},
										"type": {
											"type": "string",
											"description": "resource type identifier, 'transaction_result'"
										},
										"attributes": {
											"$ref": "#/definitions/TransactionResult"
										}
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"4xx": {
						"description": "Error with the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					},
					"5xx": {
						"description": "Error processing the request",
						"schema": {
							"type": "object",
							"properties": {
								"errors": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/APIError"
									}
								},
								"meta": {
									"type": "object",
									"description": "Extra metadata, such as API version"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"APIError": {
			"properties": {
				"id": {
					"type": "string",
					"description": "Error ID, if available"
				},
				"status": {
					"type": "string",
					"description": "HTTP status code"
				},
				"code": {
					"type": "string",
					"description": "API-specific error code, if available"
				},
				"title": {
					"type": "string",
					"description": "Short error description"
				},
				"detail": {
					"type": "string",
					"description": "Detailed error information"
				}
			}
		},
		"Contract": {
			"properties": {
				"id": {
					"type": "string",
					"description": "Contract unique identifier. In Ethereum this is the contract address."
				},
				"author": {
					"type": "string",
					"description": "Pubkey of sender on original create."
				},
				"keys": {
					"type": "array",
					"description": "Pubkeys that can access contract.",
					"items": {
						"type": "string"
					}
				},
				"abi": {
					"type": "string",
					"description": "Contract ABI"
				},
				"code": {
					"type": "string",
					"description": "Contract source code"
				},
				"code_type": {
					"type": "string",
					"description": "Type of code being submitted",
					"default": "solidity",
					"enum": [
						"solidity"
					]
				}
			}
		},
		"TransactionResult": {
			"required": [
				"id"
			],
			"properties": {
				"id": {
					"type": "string",
					"description": "ID of the newly created transaction"
				}
			}
		},
...
	}
}